+++ internal/ui/root_ui_events.go
@@ -149,6 +149,7 @@ func (m RootModel) updateUIMsg(msg tea.Msg) (RootModel, tea.Cmd) {
 	case components.SpinnerTickMsg:
 		m.chatList.TickSpinner()
 		m.chat.TickSpinner()
+		m.statusBar.TickDownloadSpinner()
 		m.updateGifLoadingSpinner()
 		m.updateVideoSpinner()
 		if m.screen == ScreenMain {